tests/wpt/mozilla/tests/mozilla/window_setInterval.html

<html>
  <head>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
    async_test(function() {
      var x = 0;
      var intervalID = setInterval(this.step_func(function() {
        x += 1;
        if (x == 2) {
          clearInterval(intervalID);
          setTimeout(this.step_func_done(function() {
            assert_equals(x, 2);
          }), 300);
        }
      }), 10);
    });
    </script>
  </body>
</html>
